Roll DEPS to latest NaCl and Chromium repos, also change to use NaCl-style scons instead of hammer.  Remove the old (and deprecated) swtoolkit in favour of using the NaCl-based site_scons packages.
Note: This CL also includes an update to make sine_synth compile w/ the latest
Pepper headers.
Review URL: http://codereview.chromium.org/6803026

git-svn-id: https://nativeclient-sdk.googlecode.com/svn/trunk/src@797 050acbb0-2703-11df-ab0a-9f3f633ae91d

+++ examples/sine_synth/sine_synth.cc
@@ -1,6 +1,6 @@
-// Copyright 2010 The Native Client Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can
-// be found in the LICENSE file.
+// Copyright (c) 2011 The Native Client Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
 
 #include <ppapi/cpp/audio.h>
 #include <ppapi/cpp/instance.h>
@@ -228,7 +228,9 @@ class SineSynthInstance : public pp::Instance {
   }
 
  private:
-  static void SineWaveCallback(void* samples, size_t buffer_size, void* data) {
+  static void SineWaveCallback(void* samples,
+                               uint32_t buffer_size,
+                               void* data) {
     SineSynthInstance* sine_synth_instance =
         reinterpret_cast<SineSynthInstance*>(data);
     const double frequency =
